Added 2 projects with some different animation algorithm
Added 2 more projects that implements a visual hint when dragging from some UI element of application.
A default Windows services was used to make such hint. Hint position updates on MouseMove event, so it's not possible to implement a smooth animation of window, because a new window position must be calculated even when user has already stopped moving the mouse cursor. But it is worth considering this method of creating such animations though